Move communication protocol to `pkg` from `internal`
