+++ pkg/client/protocol_test.go
@@ -1,4 +1,4 @@
-package internal
+package client
 
 import (
 	"bytes"
